Adjustments to AppVersion and AppVersionService
